+++ test/acceptance/todomvc.test.ts
@@ -1,5 +1,5 @@
 import { elements } from './../../lib/playright';
-import { stage, goto, element, perform } from '../../lib'
+import { stage, goto, element, perform, have } from '../../lib'
 
 describe('Todomvc', () => {
 
@@ -17,10 +17,38 @@ describe('Todomvc', () => {
         await element('#new-todo').type('a').then(perform.press('Enter'));
         await element('#new-todo').type('b').then(perform.press('Enter'));
         await element('#new-todo').type('c').then(perform.press('Enter'));
-        // await elements('#todo-list li').should();
+        await elements('#todo-list li').should(have.texts('a', 'b', 'c'));
 
         await elements('#todo-list li').first.element('.toggle').click();
+        await elements('#todo-list li').firstBy(have.text('c'))
+            .element('.toggle').click();
+        await elements('#todo-list li').by(have.cssClass('completed'))
+            .should(have.texts('a', 'c'))
+        await elements('#todo-list li').by(have.no.cssClass('completed'))
+            .should(have.texts('b'))
 
         await stage.page.waitForTimeout(4000);
     })
+
+    /**
+     * TODO: by the way, what about the following style:
+
+    fit('should complete todo', async () => {
+
+
+        await goto('http://todomvc.com/examples/emberjs');
+
+        await the('#new-todo').type('a').then(perform.press('Enter'));
+        await the('#new-todo').type('b').then(perform.press('Enter'));
+        await the('#new-todo').type('c').then(perform.press('Enter'));
+        await all('#todo-list li').should(have.texts('a', 'b', 'c'));
+
+        await all('#todo-list li').first.element('.toggle').click();
+        await all('#todo-list li').firstBy(have.text('c')).element('.toggle').click();
+        await all('#todo-list li').by(have.cssClass('completed')).should(have.texts('a', 'c'));
+        await all('#todo-list li').by(have.no.cssClass('completed')).should(have.texts('b'));
+    })
+
+     * TODO: or should we leave it to the user?
+     */
 });