+++ CHANGELOG.md
@@ -1,15 +1,35 @@
 # Changelog
 
+## TODO:
+- fix failure rendering in console log. currently it fails with:
+  Failures:
+        1) Todomvc should complete todo
+          Message:
+            [object Promise] thrown
+          Stack:
+- implement async iterator for Elements
+
 ## 0.0.next
 - TBD
 
-## 0.0.3 (to be released on 2020.08)
+## 0.0.3 (to be released on 2020.08.18)
 - moved Located to lib/experimental
-  - let's forget about it for now and stick to classic Element + Elements (Collection)
+  - let's forget about it for now and stick to classic Element + Elements (aka Collection)
 - added element and elements instead of earlier $ in main api
 - added Elemment and Elements
-- added Element#click, doubleClick, contextClick
-- added Elements#first, element(numberStartingFromOne)
+- added Element#click, doubleClick, contextClick, should
+- added to Elements: 
+    - `element(numberStartingFromOne)`
+    - `first` (alias for `element(1)`)
+    - `by` (aka `filteredBy` from SelenideJs), 
+    - `firstBy` (aka `elementBy` from SelenideJs)
+    - `should(condition)`
+- added have.
+  - text, no.text
+  - texts, no.texts
+  - cssClass, no.cssClass
+- added be.
+  - visible, not.visible
 
 ## 0.0.2 (released on 2020.08.16)
 - `goto(url)`